feat: add commands
