+++ package.json
@@ -48,6 +48,7 @@
   },
   "dependencies": {
     "@clack/prompts": "^0.6.3",
+    "cac": "^6.7.14",
     "consola": "^2.15.3",
     "picocolors": "^1.0.0",
     "tinify": "^1.7.1"
